+++ test/integration/test_boot.rb
@@ -77,7 +77,7 @@ class TestBoot < Pitchfork::IntegrationTest
       listen "#{addr}:#{port}"
       worker_processes 2
       refork_after [50, 100, 1000]
-      after_promotion do |_server, _worker|
+      after_promotion do |_server, _mold|
         raise "Oops"
       end
     RUBY